pyproject.toml
[project]
name = "ai-startup-team"
version = "0.1.0"
description = "5-agent AI founding team on Microsoft Foundry."
requires-python = ">=3.11"
dependencies = [
  "agent-framework>=1.0.0",            # Microsoft Agent Framework
  "agent-framework-foundry>=1.0.0",    # Foundry chat client + evals
  "azure-ai-projects>=1.0.0",          # Foundry Projects SDK
  "azure-ai-evaluation>=1.16.0",       # batch eval SDK
  "azure-identity>=1.17.0",
  "azure-search-documents>=11.5.0",    # vector memory
  "openai>=1.40.0",
  "anthropic>=0.40.0",                 # Claude Sonnet via Foundry model catalog
  "httpx>=0.27.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "fastapi>=0.115.0",                  # hosted agent HTTP surface
  "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.6.0", "mypy>=1.10"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# The importable package lives in `src/`; tell hatchling to ship it.
[tool.hatch.build.targets.wheel]
packages = ["src"]
